Solution to Manage Book Library

+++ Manage Book Library@@67877d61dc107486ad5c274c/code.c
@@ -2,20 +2,20 @@
 struct book{
     char title[100];
     char author[100];
-    int price;
+    float price;
 };
 int main(){
     int n;
     scanf("%d",&n);
     struct book books[n];
     for(int s=0;s<n;s++){
-        scanf("%s %s %d",&books[s].title,&books[s].author,&books[s].price);
+        scanf("%s %s %f",&books[s].title,&books[s].author,&books[s].price);
     }
-    int k;
-    scanf("%d",&k);
+    float k;
+    scanf("%f",&k);
     for(int s=0;s<n;s++){
         if(books[s].price>k){
-            printf("Books above price %d:\nTitle: %s, Author: %s, Price: %d",k,books[s].title,books[s].author,books[s].price);
+            printf("Books above price %.2f:\nTitle: %s, Author: %s, Price: %.2f",k,books[s].title,books[s].author,books[s].price);
         }
     }
 }
